Copied code and tweaked movement
i wish it didn't come to this, but for the sake of getting the game finished on time, i pasted in example code from GDquest and tweaked it to my fitting. i will definitely review it after the jam is over, though. arrays, matrixes and duplicating child instances are three complicated things that i dont have time to learn in two days, haha :)